Moves the state into the scene, renames the environments to scenes for consistency
