+++ web/stories/pages/posts/fast.stories.svelte
@@ -1,30 +1,30 @@
 <script context="module" lang="ts">
   import { Story } from '@storybook/addon-svelte-csf'
 
-  import MockedFast from './fast-mock-env.svelte'
+  import FastScene from './fast-scene.svelte'
   import { MOCKED_FAST_FEED, MOCKED_POSTS } from './mocks.js'
 
   export const meta = {
-    component: MockedFast,
+    component: FastScene,
     title: 'Pages/Fast'
   }
 </script>
 
 <Story name="Empty" parameters={{ layout: 'fullscreen' }}>
-  <MockedFast />
+  <FastScene />
 </Story>
 
 <Story name="Few posts" parameters={{ layout: 'fullscreen' }}>
-  <MockedFast posts={MOCKED_FAST_FEED.slice(0, 3)} />
+  <FastScene posts={MOCKED_FAST_FEED.slice(0, 3)} />
 </Story>
 
 <Story name="Opened post" parameters={{ layout: 'fullscreen' }}>
-  <MockedFast
+  <FastScene
     openedPost={MOCKED_POSTS[0]}
     posts={MOCKED_FAST_FEED.slice(0, 3)}
   />
 </Story>
 
 <Story name="Many posts" parameters={{ layout: 'fullscreen' }}>
-  <MockedFast posts={MOCKED_FAST_FEED} showNextButton />
+  <FastScene posts={MOCKED_FAST_FEED} showNextButton />
 </Story>
