+++ web/stories/pages/posts/slow-scene.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import {
     openedSlowPost,
-    slowPosts,
     totalSlowPages,
     totalSlowPosts
   } from '@slowreader/core'
@@ -10,15 +9,13 @@
 
   import { forceSet } from '../../../../core/lib/stores.js'
   import Slow from '../../../pages/slow.svelte'
+  import Scene from '../../scene.svelte'
 
-  export let state: SlowPostsValue = {
-    isLoading: true
-  }
+  export let state: SlowPostsValue | undefined = undefined
   export let showPagination = false
   export let post: PostValue | undefined = undefined
 
   onMount(() => {
-    forceSet(slowPosts, state)
     forceSet(totalSlowPages, showPagination ? 10 : 1)
     forceSet(totalSlowPosts, showPagination ? 1_000 : 10)
     // @ts-expect-error
@@ -26,4 +23,6 @@
   })
 </script>
 
-<Slow />
+<Scene slowState={state}>
+  <Slow />
+</Scene>
