+++ web/stories/pages/posts/slow.stories.svelte
@@ -2,33 +2,33 @@
   import { Story } from '@storybook/addon-svelte-csf'
 
   import { MOCKED_POSTS } from './mocks.js'
-  import MockedSlow from './slow-mock-env.svelte'
+  import SlowScene from './slow-scene.svelte'
 
   export const meta = {
-    component: MockedSlow,
+    component: SlowScene,
     title: 'Pages/Slow'
   }
 </script>
 
 <Story name="Loading" parameters={{ layout: 'fullscreen' }}>
-  <MockedSlow />
+  <SlowScene />
 </Story>
 
 <Story name="Empty" parameters={{ layout: 'fullscreen' }}>
-  <MockedSlow state={{ isLoading: false, list: [] }} />
+  <SlowScene state={{ isLoading: false, list: [] }} />
 </Story>
 
 <Story name="Few posts" parameters={{ layout: 'fullscreen' }}>
-  <MockedSlow state={{ isLoading: false, list: MOCKED_POSTS.slice(0, 3) }} />
+  <SlowScene state={{ isLoading: false, list: MOCKED_POSTS.slice(0, 3) }} />
 </Story>
 
 <Story name="Opened post" parameters={{ layout: 'fullscreen' }}>
-  <MockedSlow
+  <SlowScene
     post={MOCKED_POSTS[1]}
     state={{ isLoading: false, list: MOCKED_POSTS.slice(0, 3) }}
   />
 </Story>
 
 <Story name="Many posts" parameters={{ layout: 'fullscreen' }}>
-  <MockedSlow showPagination state={{ isLoading: false, list: MOCKED_POSTS }} />
+  <SlowScene showPagination state={{ isLoading: false, list: MOCKED_POSTS }} />
 </Story>
